+++ app/src/main/java/com/example/tweetssearch/ui/home/KeywordsView.kt
@@ -1,4 +1,4 @@
-package com.example.tweetssearch.ui
+package com.example.tweetssearch.ui.home
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
